backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dateutil>=2.9.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
supabase>=2.0.0
